review comment incoorpration
