+++ data/signIn.js
@@ -1,8 +1,8 @@
 module.exports = {
-  inputs: [{ label: 'Email', type: 'email' }, { label: 'Password', type: 'password' }],
+  inputs: [{ label: 'Email', type: 'email', id: 'email' }, { label: 'Password', type: 'password', id: 'password ' }],
   submitLabel: 'Login',
   primaryHeading: 'Login',
   primarySubHeading: 'Get access to your Orders, Wishlist and Recommendations',
   submitLink: '/home',
-  cart: undefined
+  cart: null
 };
