+++ resources/scss/cart.scss
@@ -1,38 +1,38 @@
 @import './_common.scss';
 
-main.cart {
+.cart {
   @extend .flex-column;
   background: #eeeeee;
 
-  .cart-header {
+  &__header {
     @extend .flex-column;
     background: black;
     padding: 10px;
-    &__text {
+    &-text {
       color: white;
     }
   }
 }
 
-.cart-item {
+.cart__item {
   @extend .flex-row;
   margin-top: 10px;
   background: $background-color-primary;
-  &__logo {
+  &-logo {
     width: 20%;
     height: 20%;
     min-width: 14rem;
   }
 }
 
-.cart-item-detail {
+.cart__item-detail {
   @extend .flex-column;
 
-  &__heading {
+  &-heading {
     font-size: 1.5rem;
     padding: 15px;
   }
-  &__info {
+  &-info {
     @extend .flex-row;
     font-size: 1rem;
     flex-direction: column;
