+++ resources/views/pages/cart.ejs
@@ -5,34 +5,32 @@
     <link rel="stylesheet" href="/resources/css/cart.css" />
   </head>
   <body class="container">
-    <header>
-      <% include ../partials/header %>
-    </header>
+    <% include ../partials/header %>
 
     <main class="cart">
-      <div class="cart-header">
-        <div class="cart-header__text">My Cart (<%= cart.count %> item)</div>
+      <div class="cart__header">
+        <div class="cart__header-text">My Cart (<%= cart.count %> item)</div>
       </div>
 
-      <% for(var i=0; i < cart.items.length; i++) { %>
-      <div class="cart-item" data-id="<%= cart.items[i].product.id %>">
+      <% for(let i=0; i < cart.items.length; i++) { %>
+      <div class="cart__item" data-id="<%= cart.items[i].product.id %>">
         <img
           src="<%= cart.items[i].product.imageURL %>"
           alt="<%= cart.items[i].product.name %> logo"
-          class="cart-item__logo"
+          class="cart__item-logo"
         />
-        <div class="cart-item-detail">
-          <h1 class="cart-item-detail__heading"><%= cart.items[i].product.name %></h1>
-          <div class="cart-item-detail__info">
-            <div class="cart-item-detail__info-count">
-              <a class="cart-item-detail__info-remove" data-id="<%= cart.items[i].product.id %>">&#45;</a>
-              <span class="cart-item-detail__info-item-count" data-id="<%= cart.items[i].product.id %>"
+        <div class="cart__item-detail">
+          <h1 class="cart__item-detail-heading"><%= cart.items[i].product.name %></h1>
+          <div class="cart__item-detail-info">
+            <div class="cart__item-detail-info-count">
+              <a class="cart__item-detail-info-remove" data-id="<%= cart.items[i].product.id %>">&#45;</a>
+              <span class="cart__item-detail-info-item-count" data-id="<%= cart.items[i].product.id %>"
                 ><%= cart.items[i].count %></span
               >
-              <a class="cart-item-detail__info-add" data-id="<%= cart.items[i].product.id %>">&#43;</a>
+              <a class="cart__item-detail-info-add" data-id="<%= cart.items[i].product.id %>">&#43;</a>
               * Rs.<%= cart.items[i].product.price %>
             </div>
-            <span class="cart-item-detail__info-total" data-id="<%= cart.items[i].product.id %>">
+            <span class="cart__item-detail-info-total" data-id="<%= cart.items[i].product.id %>">
               Rs.<%= (cart.items[i].product.price * cart.items[i].count) %>
             </span>
           </div>
@@ -41,9 +39,10 @@
       <% } %>
     </main>
 
-    <footer>
-      <% include ../partials/footer %>
-    </footer>
+    <% include ../partials/footer %>
+    <script src="/resources/js/services.js" type="text/javascript" defer></script>
+    <script src="/resources/js/headerCartController.js" type="text/javascript" defer></script>
+    <script src="/resources/js/cartItemController.js" type="text/javascript" defer></script>
     <script src="/resources/js/cart.js" type="text/javascript" defer></script>
   </body>
 </html>
