+++ resources/views/pages/index.ejs
@@ -5,9 +5,7 @@
     <link rel="stylesheet" href="/resources/css/register.css" />
   </head>
   <body>
-    <header>
-      <% include ../partials/header %>
-    </header>
+    <% include ../partials/header %>
     <main class="form-container">
       <div class="heading">
         <h1 class="heading__primary">
@@ -16,10 +14,16 @@
         <h3 class="heading__sub"><%= primarySubHeading %></h3>
       </div>
       <form class="form">
-        <% for(var i=0; i < inputs.length; i++) { %>
+        <% for(let i=0; i < inputs.length; i++) { %>
         <div class="form__input-container">
-          <input type="<%= inputs[i].type %>" class="register-form__input" required name="<%= inputs[i].label %>" />
-          <label for="name"><%= inputs[i].label %></label>
+          <input
+            type="<%= inputs[i].type %>"
+            class="register-form__input"
+            required
+            name="<%= inputs[i].label %>"
+            id="<%= inputs[i].id %>"
+          />
+          <label for="<%= inputs[i].id %>"><%= inputs[i].label %></label>
           <div class="bar"></div>
         </div>
         <% } %>
@@ -27,8 +31,6 @@
       </form>
     </main>
 
-    <footer>
-      <% include ../partials/footer %>
-    </footer>
+    <% include ../partials/footer %>
   </body>
 </html>
