+++ resources/views/partials/footer.ejs
@@ -1,1 +1,1 @@
-<p class="app-footer">Copyright &copy; 2011-2019 Sabka Bazaar Grocery Supplies Pvt Ltd</p>
+<footer><p class="app-footer">Copyright &copy; 2011-2019 Sabka Bazaar Grocery Supplies Pvt Ltd</p></footer>
